Feature/component improvement

+++ resources/assets/js/components/Organization/Organization.vue
@@ -3,7 +3,7 @@
   <div class="box">
     <div>
       <div class="search-box">
-        <input v-model="organizationId" class="input search-input" type="text" placeholder="请输入你要查看的组织">
+        <input v-model="searchKey" class="input search-input" type="text" placeholder="请输入你要查看的组织">
         <button @click="searchOrganization()" class="button" type="button" name="button">查找组织</button>
       </div>
         <button @click="addOrganization()" class="button add-role-button" type="button" name="button">添加组织</button>
@@ -20,15 +20,16 @@
           <th>当前容量</th>
           <th>创建时间</th>
           <th>更新时间</th>
-          <th>操作</th>
+          <th>组织操作</th>
+          <th>成员操作</th>
         </tr>
       </thead>
       <tbody>
-        <tr v-for="item in OrganizationData">
+        <tr v-for="(item,index) in organizationData">
           <td>{{ item.id }}</td>
           <td>{{ item.name }}</td>
           <td>{{ item.creator_id }}</td>
-          <td> {{ item.description }}</td>
+          <td>{{ item.description }}</td>
           <td>{{ item.max }}</td>
           <td>{{ item.current }}</td>
           <td>{{ item.created_at }}</td>
@@ -36,14 +37,30 @@
           <td>
             <button @click="deleteOrganization(index)" class="button" type="button" name="button">删除组织</button>
             <button @click="editOrganization(index)"  class="button" type="button" name="button">编辑组织</button>
-            <button  class="button" type="button" name="button">查看成员</button>
+          </td>
+          <td>
+            <button @click="showMember(index)"  class="button" type="button" name="button">查看成员</button>
           </td>
         </tr>
       </tbody>
     </table>
 
-    <add-organization ref="addOrganization"></add-organization>
-    <edit-organization ref="editOrganization"  :edit-data="editData"></edit-organization>
+    <add-organization ref="addOrganization"
+                      v-on:getOrganization="getOrganization"
+    ></add-organization>
+
+    <edit-organization ref="editOrganization"
+                       v-on:getOrganization="getOrganization"
+                       v-bind:edit-data="editData"
+    ></edit-organization>
+
+    <member ref="member"
+            v-bind:current-organization-data="currentOrganizationData"
+    ></member>
+
+    <pagination v-bind:pagination-data="paginationData"
+                v-model="data"
+    ></pagination>
 
   </div>
 </template>
@@ -51,32 +68,27 @@
 <script>
 import AddOrganization from './AddOrganization'
 import EditOrganization from './EditOrganization'
+import Pagination from './../Pagination'
+import Member from './../Member/Member'
 
 export default {
   data() {
     return {
       token: null,
-      "OrganizationData": [
-          {
-              "id": 2,
-              "name": "1班1111",
-              "creator_id": "1",
-              "": "职教师资1班",
-              "max": "234",
-              "current": "1",
-              "created_at": "2018-01-13 08:04:13",
-              "updated_at": "2018-01-13 08:08:55"
-          }
-      ],
+      organizationData: null,
       isShowModal: false,
-      // OrganizationData: null,
-      organizationId: null,
+      searchKey: null,
       editData: null,  // 当前编辑的组织数据
+      paginationData: null,
+      data: null,
+      currentOrganizationData: null,
     }
   },
   components: {
     AddOrganization,
     EditOrganization,
+    Pagination,
+    Member,
   },
   methods: {
     addOrganization: function() {
@@ -85,6 +97,7 @@ export default {
     },
     editOrganization: function (index) {
       const that = this;
+      that.editData = that.organizationData[index];
       that.$refs.editOrganization.switchModal();
     },
     getOrganization: function () {
@@ -97,7 +110,8 @@ export default {
           'Authorization': that.token
         }
       }).then(res => {
-        that.permissionData = res.data.data;
+        that.organizationData = res.data.data;
+        that.paginationData = res.data.links;
       }).catch(err => {
         console.log(err)
       })
@@ -106,34 +120,49 @@ export default {
       const that = this;
       axios({
         method: 'get',
-        url: `${this.GLOBAL.localDomain}/api/v1/organizations/${that.organizationId}`,
+        url: `${this.GLOBAL.localDomain}/api/v1/organizations/${that.searchKey}`,
         headers: {
           'Accept': 'application/json',
           'Authorization': that.token
         }
       }).then(res => {
-        that.permissionData = [];
-        that.permissionData.push(res.data.data);
+        that.organizationData = [];
+        that.organizationData.push(res.data.data);
       }).catch(err => {
         console.log(err)
       })
     },
-    // 删除组织 ??删除需要的参数
     deleteOrganization: function (index) {
       const that = this;
       let id = that.organizationData[index]['id'];
-      axios({
-        method: 'delete',
-        url: `${this.GLOBAL.localDomain}/api/v1/organizations/${id}`,
-        headers: {
-          'Accept': 'application/json',
-          'Authorization': that.token
-        }
-      }).then(res => {
-      }).catch(err => {
-        console.log(err)
-      })
-    }
+      let prompt = confirm("确认删除该组织吗？");
+      if (prompt) {
+        axios({
+          method: 'delete',
+          url: `${this.GLOBAL.localDomain}/api/v1/organizations/${id}`,
+          headers: {
+            'Accept': 'application/json',
+            'Authorization': that.token
+          }
+        }).then(res => {
+          alert('删除成功');
+          that.getOrganization();
+        }).catch(err => {
+          alert('删除失败');
+          console.log(err)
+        })
+      }
+    },
+    showMember: function (index) {
+      const that = this;
+      that.currentOrganizationData = that.organizationData[index];
+      that.$refs.member.switchModal();
+    },
+    addMember: function (index) {
+      const that = this;
+      that.currentOrganizationData = that.organizationData[index];
+      that.$refs.addMember.switchModal();
+    },
   },
   computed: {
     isShowCreateOrganization() {
@@ -151,6 +180,11 @@ export default {
     this.getOrganization();
   },
   watch: {
+    data:function (value, oldValue) {
+      const that = this;
+      that.organizationData = value.data;
+      that.paginationData = value.links;
+    }
   }
 }
 </script>
